generalizing code to enable virtual fields
