+++ src/exanb/analytics/CMakeLists.txt
@@ -1,0 +1,3 @@
+set(exanbAnalytics_LINK_LIBRARIES exanbGridCellParticles exanbMPI)
+xstamp_add_plugin(exanbAnalytics ${CMAKE_CURRENT_SOURCE_DIR})
+
