+++ src/exanb/analytics/grid_particle_slicing.cpp
@@ -1,0 +1,80 @@
+#include <exanb/core/operator.h>
+#include <exanb/core/operator_slot.h>
+#include <exanb/core/operator_factory.h>
+#include <exanb/core/grid.h>
+#include <exanb/core/domain.h>
+#include <exanb/core/basic_types.h>
+#include <exanb/core/make_grid_variant_operator.h>
+#include <exanb/grid_cell_particles/grid_particle_field_accessor.h>
+#include <onika/plot1d.h>
+#include <exanb/analytics/grid_particle_slicing.h>
+
+#include <regex>
+#include <type_traits>
+#include <mpi.h>
+
+//#include <exanb/compute/math_functors.h>
+// allow field combiner to be processed as standard field
+//ONIKA_DECLARE_FIELD_COMBINER( exanb, VelocityNorm2Combiner , vnorm2 , exanb::Vec3Norm2Functor , exanb::field::_vx , exanb::field::_vy , exanb::field::_vz )
+
+namespace exanb
+{
+
+  template< class GridT >
+  class GridParticleSlicing : public OperatorNode
+  {
+    using StringList = std::vector<std::string>;
+    using Plot1DSet = onika::Plot1DSet;
+    using StringMap = std::map<std::string,std::string>;
+    static constexpr FieldSet<field::_rx,field::_ry,field::_rz> reduce_field_set {};
+    
+    ADD_SLOT( MPI_Comm   , mpi        , INPUT , MPI_COMM_WORLD );
+    ADD_SLOT( GridT      , grid       , INPUT , REQUIRED );
+    ADD_SLOT( Domain     , domain     , INPUT , REQUIRED );
+    ADD_SLOT( double     , thickness  , INPUT , 1.0 );
+    ADD_SLOT( Vec3d      , direction  , INPUT , Vec3d{1,0,0} );
+    ADD_SLOT( StringList , fields     , INPUT , StringList({".*"}) , DocString{"List of regular expressions to select fields to slice"} );
+    ADD_SLOT( StringMap  , plot_names , INPUT , StringMap{} , DocString{"map field names to output plot names"} );
+    ADD_SLOT( Plot1DSet  , plots      , INPUT_OUTPUT );
+
+  public:
+
+    inline void execute ()  override final
+    {    
+      execute_on_field_set(grid->field_set);
+    }
+    
+  private:
+    template<class... GridFields>
+    inline void execute_on_fields( const GridFields& ... grid_fields) 
+    {
+      const auto& flist = *fields;
+      auto field_selector = [&flist] ( const std::string& name ) -> bool { for(const auto& f:flist) if( std::regex_match(name,std::regex(f)) ) return true; return false; } ;
+      auto pecfunc = [op=this]() { return op->parallel_execution_context(); };
+      using ParticleAcessor = GridParticleFieldAccessor<typename GridT::CellParticles *>;
+      ParticleAcessor pacc = {grid->cells()};
+      slice_grid_particles( *mpi, *grid, domain->xform(), *direction, *thickness, *plots, pacc, field_selector, pecfunc, grid_fields... );
+    }
+
+    template<class... fid>
+    inline void execute_on_field_set( FieldSet<fid...> ) 
+    {
+      execute_on_fields( onika::soatl::FieldId<fid>{} ... );
+    }
+
+    // -----------------------------------------------
+    // -----------------------------------------------
+    inline std::string documentation() const override final
+    {
+      return R"EOF(create 1D Plots from particles fields, averaging those fields by slices of given orientation and thickness)EOF";
+    }    
+
+  };
+
+  // === register factories ===
+  CONSTRUCTOR_FUNCTION
+  {
+    OperatorNodeFactory::instance()->register_factory("grid_particle_slicing", make_grid_variant_operator< GridParticleSlicing > );
+  }
+
+}
